third_party/WebKit/LayoutTests/external/wpt/html/webappapis/system-state-and-capabilities/the-navigator-object/navigatorlanguage-expected.txt

This is a testharness.js-based test.
FAIL NavigatorLanguage: the most preferred language is the one returned by navigator.languageassert_equals: navigator.languages is the most preferred language first expected (string) "en-US" but got (undefined) undefined
Harness: the test ran to completion.
